backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
pydantic==2.5.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
# uuid是Python标准库，无需额外安装
